requirements.txt
# ============================================================
# Neo Eco Cleaning — AI Marketing Dashboard
# ============================================================

# --- LLM (Gemini Pro) ---
google-genai>=1.0.0
httpx>=0.27.0

# --- Web Dashboard ---
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.9

# --- Email ---
aiosmtplib>=3.0.0
jinja2>=3.1.0

# --- Data ---
pandas>=2.2.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
openpyxl>=3.1.0

# --- Apollo API ---
requests>=2.31.0

# --- CLI ---
rich>=13.7.0
click>=8.1.0

# --- Utilities ---
python-dateutil>=2.9.0
